Rename AppSettings to EnvSettings
This is an attempt to reduce the confusion between "config" and
"app settings".  The app settings are those values that are read from
the environment (either env variables or .env file), so renaming it to
"env settings" hopefully makes it clearer.
